[modbus] ModbusRegisterArray abstraction removed
