+++ bundles/org.openhab.io.transport.modbus/src/test/java/org/openhab/io/transport/modbus/test/BitUtilitiesExtractStringFromRegistersTest.java
@@ -32,7 +32,7 @@ import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import org.junit.runners.Parameterized.Parameters;
 import org.openhab.io.transport.modbus.ModbusRegister;
-import org.openhab.io.transport.modbus.BasicModbusRegisterArray;
+import org.openhab.io.transport.modbus.ModbusRegisterArray;
 import org.openhab.io.transport.modbus.ModbusBitUtilities;
 import org.openhab.io.transport.modbus.ModbusRegister;
 import org.openhab.io.transport.modbus.ModbusRegisterArray;
@@ -63,7 +63,7 @@ public class BitUtilitiesExtractStringFromRegistersTest {
 
     private static ModbusRegisterArray shortArrayToRegisterArray(int... arr) {
         ModbusRegister[] tmp = new ModbusRegister[0];
-        return new BasicModbusRegisterArray(IntStream.of(arr).mapToObj(val -> {
+        return new ModbusRegisterArray(IntStream.of(arr).mapToObj(val -> {
             ByteBuffer buffer = ByteBuffer.allocate(2);
             buffer.putShort((short) val);
             return new ModbusRegister(buffer.get(0), buffer.get(1));
